#230 - Permite o component List tenho componentes filhos
