+++ package.json
@@ -25,29 +25,29 @@
     "pre-commit": "concurrently -r \"yarn lint\" \"yarn ultra -r pre-commit\""
   },
   "devDependencies": {
-    "@babel/core": "7.14",
-    "@types/faker": "5.5.7",
-    "@typescript-eslint/eslint-plugin": "4.28.3",
-    "@typescript-eslint/parser": "4.28.3",
-    "concurrently": "6.2.0",
-    "eslint": "7.30.0",
+    "@babel/core": "7.15",
+    "@types/faker": "5.5.8",
+    "@typescript-eslint/eslint-plugin": "4.31.0",
+    "@typescript-eslint/parser": "4.31.0",
+    "concurrently": "6.2.1",
+    "eslint": "7.32.0",
     "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-eslint-plugin": "3.4.0",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-mdx": "1.14.0",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-react": "7.24.0",
+    "eslint-plugin-eslint-plugin": "3.5.3",
+    "eslint-plugin-import": "2.24.2",
+    "eslint-plugin-mdx": "1.15.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-react": "7.25.1",
     "eslint-plugin-react-hooks": "4.2.0",
-    "eslint-plugin-sonarjs": "0.9.1",
-    "eslint-plugin-unused-imports": "1.1.2",
+    "eslint-plugin-sonarjs": "0.10.0",
+    "eslint-plugin-unused-imports": "1.1.4",
     "faker": "5.5.3",
-    "husky": "7.0.1",
-    "inquirer": "8.1.1",
-    "ora": "5.4.1",
+    "husky": "7.0.2",
+    "inquirer": "8.1.2",
+    "ora": "5.4",
     "patch-package": "*",
-    "prettier": "2.3.2",
+    "prettier": "2.4.0",
     "semver": "7.3.5",
-    "typescript": "4.3",
+    "typescript": "4.4",
     "ultra-runner": "3.10.5"
   },
   "resolutions": {
