+++ src/dev/package.json
@@ -13,8 +13,8 @@
     "react-scripts": "4.0.3"
   },
   "devDependencies": {
-    "@types/jest": "26.0.23",
-    "@types/node": "15.12.1",
+    "@types/jest": "27.0.1",
+    "@types/node": "16.9.1",
     "@types/react": "^16",
     "@types/react-dom": "^16"
   },
