+++ src/packages/experimental/package.json
@@ -26,7 +26,7 @@
     "@eduzz/houston-hooks": "0.31.3",
     "@eduzz/houston-icons": "0.31.3",
     "@emotion/css": "^11.1.3",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@material-ui/lab": "^4.0.0-alpha.60",
     "clsx": "^1.1.1",
     "tslib": "^2"
   },
