--- src/packages/ui-components/List/Title/index.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-import { TypographyProps } from '@material-ui/core/Typography';
-
-type ListTitleProps = 'id' | 'className';
-
-export interface IListTitleProps extends Pick<TypographyProps, ListTitleProps> {
-  children: React.ReactNode;
-}
-
-export default (() => null) as React.FC<IListTitleProps>;
