--- src/packages/ui-components/List/internals/Subtitle/index.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-
-import Typography from '../../../Typography';
-import { IListTitleProps } from '../../Title';
-
-interface IProps extends IListTitleProps {}
-
-const Subtitle = ({ children, ...props }: IProps) => (
-  <Typography {...props} size='small' fontWeight='regular' lineHeight='compact'>
-    {children}
-  </Typography>
-);
-
-export default React.memo(Subtitle);
