+++ src/packages/ui-components/package.json
@@ -28,13 +28,13 @@
     "@emotion/css": "^11.1.3",
     "@material-ui/core": "^4",
     "@material-ui/icons": "^4",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@material-ui/lab": "^4.0.0-alpha.60",
     "clsx": "^1.1.1",
-    "date-fns": "^2.22.1",
+    "date-fns": "^2.23.0",
     "is-mobile": "^3.0.0",
     "lodash": "^4",
-    "react-colorful": "^5.3.0",
-    "react-date-picker": "^8.2.0",
+    "react-colorful": "^5.4.0",
+    "react-date-picker": "^8.3.2",
     "react-toastify": "^7.0.4",
     "tslib": "^2",
     "use-context-selector": "^1.3.7"
